Rework Matches into a shared matchweek list
